# apps/api/openapi/taskflow.yaml
openapi: 3.0.3
info:
  title: TaskFlow Scheduler API
  description: The main service for managing scheduled tasks in the TaskFlow system. This service handles authentication, request validation, task scheduling, and persistence.
  version: 1.0.0
  contact:
    name: TaskFlow API Support
    url: https://github.com/hhace/taskflow
    email: support@taskflow.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8081
    description: Local development server
  - url: https://api.taskflow.com
    description: Production server

paths:
  /api/v1/tasks:
    get:
      tags:
        - Tasks
      summary: List tasks
      description: Retrieve a list of tasks with optional status filtering
      operationId: getTasks
      parameters:
        - name: status
          in: query
          description: Filter tasks by status
          required: false
          schema:
            type: string
            enum: [pending, running, completed, failed]
          example: pending
      responses:
        '200':
          description: Tasks retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Tasks
      summary: Create a new task
      description: Create a new scheduled task with cron expression and command
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
            example:
              schedule: "0 */5 * * * *"
              command: ["echo", "hello world"]
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/tasks/{id}:
    get:
      tags:
        - Tasks
      summary: Get a task by ID
      description: Retrieve a specific task by its unique identifier
      operationId: getTask
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID (UUID)
          schema:
            type: string
            format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Invalid task ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    CreateTaskRequest:
      type: object
      required:
        - schedule
        - command
      properties:
        schedule:
          type: string
          description: Cron expression for task scheduling
          example: "0 */5 * * * *"
          pattern: '^(\*|([0-9]|1[0-9]|2[0-9]|3[0-9]|4[0-9]|5[0-9])|\*\/([0-9]|1[0-9]|2[0-9]|3[0-9]|4[0-9]|5[0-9])) (\*|([0-9]|1[0-9]|2[0-9]|3[0-9]|4[0-9]|5[0-9])|\*\/([0-9]|1[0-9]|2[0-9]|3[0-9]|4[0-9]|5[0-9])) (\*|([0-9]|1[0-9]|2[0-3])|\*\/([0-9]|1[0-9]|2[0-3])) (\*|([1-9]|1[0-9]|2[0-9]|3[0-1])|\*\/([1-9]|1[0-9]|2[0-9]|3[0-1])) (\*|([1-9]|1[0-2])|\*\/([1-9]|1[0-2])) (\*|([0-6])|\*\/([0-6]))$'
        command:
          type: array
          items:
            type: string
          description: Command and arguments to execute
          example: ["echo", "hello world"]
          minItems: 1

    TaskResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique task identifier
          example: "123e4567-e89b-12d3-a456-426614174000"
        schedule:
          type: string
          description: Cron expression for task scheduling
          example: "0 */5 * * * *"
        command:
          type: array
          items:
            type: string
          description: Command and arguments to execute
          example: ["echo", "hello world"]
        status:
          type: string
          enum: [pending, running, completed, failed]
          description: Current task status
          example: pending
        created_at:
          type: string
          format: date-time
          description: Task creation timestamp
          example: "2023-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Task last update timestamp
          example: "2023-01-01T00:00:00Z"

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, unhealthy]
          description: Service health status
          example: healthy
        service:
          type: string
          description: Service name
          example: scheduler
        database:
          type: string
          description: Database connection status
          example: connected
        error:
          type: string
          description: Error message if unhealthy
          example: "database connection failed"

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid task ID"

    TaskListResponse:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskResponse'
          description: List of tasks
        count:
          type: integer
          description: Number of tasks returned
          example: 10

tags:
  - name: Health
    description: Health check operations
  - name: Tasks
    description: Operations about tasks